requirements.txt
asgiref==3.11.1
attrs==25.4.0
dj-database-url==3.1.0
Django==6.0.2
django-filter==25.2
django-cors-headers==4.4.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
drf-yasg==1.21.14
gunicorn==25.0.3
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.11.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0

# Realtime chat (Django Channels)
channels==4.2.0
channels-redis==4.2.0

# ASGI server (production)
daphne==4.1.2

# Optional (recommended) Redis cache for presence/unread tracking
django-redis==5.4.0
redis==5.2.1

whitenoise==6.7.0
